[py] break up the monolithic bindings compile unit
